requirements.txt
# K-Quant v3.5 PRO
python-telegram-bot[job-queue]>=20.7
apscheduler>=3.10
requests>=2.31
beautifulsoup4>=4.12
aiohttp>=3.9
numpy>=1.24,<2.0
pandas>=2.0
scikit-learn>=1.3
lightgbm>=4.0
anthropic>=0.18
finance-datareader>=0.9
opendartreader>=0.2
openpyxl>=3.1
google-api-python-client>=2.0
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.0
httpx>=0.25
pyyaml>=6.0
python-dotenv>=1.0
pyarrow>=14.0
yfinance>=0.2.30
pytest>=7.0
pytest-asyncio>=0.21
